+++ src/SpecRunnerConfig.ts
@@ -81,6 +81,10 @@ export class SpecRunnerConfig {
     return this.getBooleanConfig('ruby-spec-runner.rspecRunButton', true);
   }
 
+  get rspecDebugButton(): boolean {
+    return this.getBooleanConfig('ruby-spec-runner.rspecDebugButton', true);
+  }
+
   get rspecRunAllFailedButton(): boolean {
     return this.getBooleanConfig('ruby-spec-runner.rspecRunAllFailedButton', false);
   }