+++ src/extension.ts
@@ -3,7 +3,7 @@ import { SpecRunnerConfig } from './SpecRunnerConfig';
 import SpecResultInterpreter from './rspec/SpecResultInterpreter';
 import SpecResultPresenter from './SpecResultPresenter';
 import { MinitestRunner, MinitestRunnerCodeLensProvider, MinitestRunnerButton, MinitestResultInterpreter } from './minitest';
-import { SpecRunner, SpecRunnerCodeLensProvider, FailedSpecRunnerButton, SpecRunnerButton } from './rspec';
+import { SpecRunner, SpecRunnerCodeLensProvider, FailedSpecRunnerButton, SpecRunnerButton, SpecDebugButton } from './rspec';
 import { RunRspecOrMinitestArg } from './types';
 
 const buildFileRunnerHandler = (minitestRunner: MinitestRunner, specRunner: SpecRunner, debugging: boolean) => async (args: any) => {
@@ -98,6 +98,7 @@ export function activate(context: vscode.ExtensionContext) {
 
   const failedSpecRunnerButton = new FailedSpecRunnerButton(vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 1), config);
   const specRunnerButton = new SpecRunnerButton(vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 2), config);
+  const specDebugButton = new SpecDebugButton(vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 2), config);
   const minitestRunnerButton = new MinitestRunnerButton(vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 2), config);
 
   context.subscriptions.push(runRspecOrMinitestFile);
@@ -109,11 +110,13 @@ export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(specCodeLensProviderDisposable);
   context.subscriptions.push(minitestCodeLensProviderDisposable);
   context.subscriptions.push(specRunnerButton.button);
+  context.subscriptions.push(specDebugButton.button);
   context.subscriptions.push(failedSpecRunnerButton.button);
   context.subscriptions.push(minitestRunnerButton.button);
   context.subscriptions.push(vscode.window.onDidChangeActiveTextEditor((editor) => {
     failedSpecRunnerButton.update(editor);
     specRunnerButton.update(editor);
+    specDebugButton.update(editor);
     minitestRunnerButton.update(editor);
     resultPresenter.update();
   }));
@@ -124,6 +127,7 @@ export function activate(context: vscode.ExtensionContext) {
 
   failedSpecRunnerButton.update();
   specRunnerButton.update();
+  specDebugButton.update();
   minitestRunnerButton.update();
 }