+++ src/rspec/index.ts
@@ -1,6 +1,7 @@
 export * from './FailedSpecRunnerButton';
 export * from './SpecRunner';
 export * from './SpecRunnerButton';
+export * from './SpecDebugButton';
 export * from './SpecRunnerCodeLensProvider';
 export * from './SpecResultInterpreter';
 export * from './SpecParser';